TODO: Semaphoe|Lock. Довести все машины до финиша

+++ lesson_7/src/main/java/name/imh0t3mp/course/geekbrains/Road.java
@@ -14,7 +14,7 @@ public class Road extends Stage {
     public void go(Car c) {
         try {
             System.out.println(c.getName() + " начал этап: " + description);
-            System.out.println(c.getName() + " расчётное время прохождения этапа:" + ((float) ((length * 1000) / c.getSpeed())));
+            System.out.println(c.getName() + " расчётное время прохождения этапа:" + ((length * 1000) / c.getSpeed()));
             Thread.sleep(length * 1000 / c.getSpeed());
             System.out.println(c.getName() + " закончил этап: " + description);
         } catch (InterruptedException e) {
